server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7

# Core dependencies
numpy>=1.20.0
Pillow>=9.0.0
torch>=1.10.0
torchvision>=0.11.0

# ML dependencies
scikit-learn>=1.0.0
matplotlib>=3.5.0

# Utils
requests>=2.25.0
python-dotenv>=0.19.0

# Dev dependencies
pytest>=6.2.5